+++ src/pages/Shared/ModalSelecteProduct.tsx
@@ -197,7 +197,7 @@ export const ModalSelectedProduct = ({
   const handleClearSearch = () => {
     setKeyword("");
     setProdGroup("");
-    setLocation("");
+    //setLocation("");
     setSelectedStrategy("");
     setSelectedBrand("");
     setSelectedProd(prodList);
@@ -344,7 +344,7 @@ export const ModalSelectedProduct = ({
         <Form layout='vertical'>
           <Row gutter={16}>
             <Col span={9}>
-              <Form.Item label='ค้นหาสินค้า' name='searchText'>
+              <Form.Item label='ค้นหาสินค้า'>
                 <Input
                   placeholder='ค้นหาชื่อสินค้าหรือรหัสสินค้า...'
                   suffix={<SearchOutlined />}
@@ -357,7 +357,7 @@ export const ModalSelectedProduct = ({
               </Form.Item>
             </Col>
             <Col span={4}>
-              <Form.Item label='Product Group' name='productGroup'>
+              <Form.Item label='Product Group'>
                 <Select
                   data={[
                     {
@@ -378,10 +378,9 @@ export const ModalSelectedProduct = ({
                 />
               </Form.Item>
             </Col>
-
             {company === "ICPL" && (
               <Col span={4}>
-                <Form.Item label='Strategy Group' name='strategyGroup'>
+                <Form.Item label='Strategy Group'>
                   <Select
                     data={[
                       {
@@ -405,7 +404,7 @@ export const ModalSelectedProduct = ({
             )}
             {brand?.length && (
               <Col span={4}>
-                <Form.Item label='ยี่ห้อ' name='brand'>
+                <Form.Item label='ยี่ห้อ'>
                   <Select
                     data={[
                       {
@@ -427,7 +426,7 @@ export const ModalSelectedProduct = ({
                 </Form.Item>
               </Col>
             )}
-            {company === "ICPI" && (
+            {/* {company === "ICPI" && (
               <Col span={4}>
                 <Form.Item label='Location' name='brand'>
                   <Select
@@ -450,7 +449,7 @@ export const ModalSelectedProduct = ({
                   />
                 </Form.Item>
               </Col>
-            )}
+            )} */}
             <Col span={3} style={{ paddingTop: "22px" }}>
               <Button title='ล้างการค้นหา' typeButton='primary-light' onClick={handleClearSearch} />
             </Col>